Make animation pause smartly when not animating
Ensure first patch of circles paints. When a suitable change occurs and
animating is off, make sure another paint occurs. Noteably, when canvas
zoom or pan changes and when numCircles or animationDuration changes.

The paused paint is a frozen paint that never changes as long as
animating stays off, using the same ellapsed time.