Better handle dragging when mouse leaves canvas (not perfect)
Dragging is stopped if the mouse re-enters the canvas with the left mouse
button up. If the left mouse button down (even if it was lifted and re-
clicked down outside the canvas), dragging is continued. It would be nice
to detect when the left mouse button is lifted outside the canvas so that
re-downs are not handled as though the button had stayed down continuously.