fix (or partial fix) for pointOnEllipseInDirectionOfAnotherPoint
The point at parametric-equation angle alpha on an ellipse is not for angle
alpha from the positive x-axis (or major axis). I can't quite think of
how to describe it, but I can describe the relationship between what I
thought it was and what it actually is. Consider a point
  e(alpha) = (Rx*cos(alpha), Ry*sin(alpha))
on an ellipse at the origin with zero rotation. Then the "true" angle to
the point is
  alpha2 = arctan(Ry/Rx * tan(alpha))
(deal with the normal issues of domain of tan/arctan).
So we just have to deal with this translation/equation. I think I have
done so for the case of non-rotated ellipses. Need to deal with rotation
correctly, now.